Add rule_dir_stats utility for analyzing rule directories
The utils/rule_dir_stats.py function analyzes the output
from utils/rule_dir_json.py and provides several useful
functions for analyzing the state of the content of rule
directories. Among these are:

 - Listing rules missing checks and fixes
 - Listing rules with two or more of a check or fix
 - Listing rules which differ in prodtypes between the rule.yml and check/fix content
 - Listing rules which have different prodtypes within a check/fix as their name
 - Introspecting a single rule within the JSON output
 - Look for rules which are unassociated to any product

These can be filtered by a list of products and/or by a list of rule_ids,
with the output being limited to checking only OVALs or fixes. Summary
statistics can also be provided.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>