no_shelllogin_for_systemaccounts: make sure variable used as object is not empty
The test collects all UIDs with a shell (except root), assign them to a
variable, make some calculations and compare to state.
If no user has a shell, there will be no UIDs collected, rendering the
calculations empty, and leading to an empty object set to be compared to
states.

This change injects a valid UID to avoid this empty object set.